refactor building of RPI object
